[pvr.tvh] introduce a struct for storing all addon settings
This struct is created when the addon is started and stored as a const
reference in CTvheadend. This way we avoid having to lock the global
mutex whenever a setting value has to be read. It was generally not
needed before either since the only time a setting could change was
right before the addon was restarted or when the timeout parameters
where changed (which didn't require a restart). This removes that
uncertainty.